+++ src/modeltest/Heuristics.cpp
@@ -1,11 +1,11 @@
 #include "Heuristics.hpp"
-#include "FreerateHeuristic.hpp"
-#include "ModelDefinitions.hpp"
-#include "RHASHeuristic.hpp"
 
 
-
-Heuristics::Heuristics(size_t partition_count, HeuristicSelection selection, const std::vector<RateHeterogeneityDescriptor> selected_rhas, const SubstitutionModelDescriptor &reference_matrix, unsigned int min_freerate_categories, unsigned int max_freerate_categories, double significant_ic_difference, RHASHeuristicMode rhas_mode)
+Heuristics::Heuristics(size_t partition_count, HeuristicSelection selection,
+                       const std::vector<RateHeterogeneityDescriptor> selected_rhas,
+                       const SubstitutionModelDescriptor &reference_matrix,
+                       unsigned int min_freerate_categories, unsigned int max_freerate_categories,
+                       double significant_ic_difference, RHASHeuristicMode rhas_mode)
     : selection{selection},
       reference_matrix{reference_matrix},
       selected_rhas{selected_rhas},
@@ -16,7 +16,6 @@ Heuristics::Heuristics(size_t partition_count, HeuristicSelection selection, con
       invariant_freerate_heuristics{enabled(HeuristicType::FREERATE) ? partition_count : 0,
                 FreerateHeuristic(min_freerate_categories, max_freerate_categories, RateHeterogeneityType::INVARIANT_FREE_RATE)}
 {
-
     for (auto p = 0UL; p < rhas_heuristics.size(); ++p)
     {
         rhas_heuristics.at(p).set_partition_index(p);
@@ -24,7 +23,7 @@ Heuristics::Heuristics(size_t partition_count, HeuristicSelection selection, con
 
 }
 
-void Heuristics::update(unsigned int partition, const ModelDescriptor &candidate_model, double score)
+void Heuristics::update(size_t partition, const ModelDescriptor &candidate_model, double score)
 {
     if (enabled(HeuristicType::FREERATE)) {
         freerate_heuristics.at(partition).update(candidate_model, score);
@@ -41,18 +40,18 @@ void Heuristics::update(unsigned int partition, const ModelDescriptor &candidate
         const int freerate_optimal_category_count = freerate_heuristics.at(partition).optimal_category_count(reference_matrix);
         if (freerate_optimal_category_count > 0) {
             rhas_heuristics.at(partition).set_optimal_category_count(RateHeterogeneityType::FREE_RATE, 
-                    freerate_optimal_category_count);
+                    (unsigned int) freerate_optimal_category_count);
         }
 
         const int inv_freerate_optimal_category_count = invariant_freerate_heuristics.at(partition).optimal_category_count(reference_matrix);
         if (inv_freerate_optimal_category_count > 0) {
             rhas_heuristics.at(partition).set_optimal_category_count(RateHeterogeneityType::INVARIANT_FREE_RATE, 
-                    inv_freerate_optimal_category_count);
+                    (unsigned int) inv_freerate_optimal_category_count);
         }
     }
 }
 
-bool Heuristics::can_skip(unsigned int partition, const ModelDescriptor &candidate_model) const
+bool Heuristics::can_skip(size_t partition, const ModelDescriptor &candidate_model) const
 {
     if (enabled(HeuristicType::FREERATE) &&
             freerate_heuristics.at(partition).can_skip(candidate_model)) {
@@ -67,11 +66,13 @@ bool Heuristics::can_skip(unsigned int partition, const ModelDescriptor &candida
     return false;
 }
 
-bool Heuristics::enabled(const HeuristicType &heuristic) const {
+bool Heuristics::enabled(const HeuristicType &heuristic) const
+{
     return selection.find(heuristic) != selection.cend();
 }
 
-bool Heuristics::evaluation_essential(unsigned int partition, const ModelDescriptor &c) const {
+bool Heuristics::evaluation_essential(size_t partition, const ModelDescriptor &c) const
+{
     bool essential = false;
 
     // Need to compute all RHAS variants of the reference substition matrix
